--- appeal-api/src/main/java/com/appeal/api/portfolio/domain/templateone/TemplateOneCertificate.java
@@ -1,33 +1,0 @@
-package com.appeal.api.portfolio.domain.templateone;
-
-import com.appeal.api.portfolio.dto.template.templateone.TemplateOneCertificateDto;
-import lombok.*;
-
-import javax.persistence.*;
-
-@Builder
-@AllArgsConstructor
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Entity
-public class TemplateOneCertificate {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "TEMPLATE_ID")
-    private TemplateOne templateOne;
-
-    private String date;
-    private String title;
-    private String origin;
-
-    public static TemplateOneCertificate createTemplateOneCertificate(TemplateOneCertificateDto certificate, TemplateOne templateOne) {
-        TemplateOneCertificate templateOneCertificate = new TemplateOneCertificate();
-        templateOneCertificate.date = certificate.getDate();
-        templateOneCertificate.title = certificate.getTitle();
-        templateOneCertificate.origin = certificate.getOrigin();
-        templateOneCertificate.templateOne = templateOne;
-        return null;
-    }
-}
