+++ appeal-api/src/main/java/com/appeal/api/portfolio/dto/PortfolioFileDto.java
@@ -1,0 +1,13 @@
+package com.appeal.api.portfolio.dto;
+
+import lombok.*;
+import org.springframework.web.multipart.MultipartFile;
+
+@Getter @Setter
+public class PortfolioFileDto {
+    private MultipartFile thumbnail;
+    private String title;
+    private String skill;
+    private String name;
+    private String intro;
+}
