--- appeal-api/src/main/java/com/appeal/api/portfolio/dto/portfolio/PortfolioFileDto.java
@@ -1,13 +1,0 @@
-package com.appeal.api.portfolio.dto.portfolio;
-
-import lombok.*;
-import org.springframework.web.multipart.MultipartFile;
-
-@Getter @Setter
-public class PortfolioFileDto {
-    private MultipartFile thumbnail;
-    private String title;
-    private String skill;
-    private String name;
-    private String intro;
-}
