--- appeal-api/src/main/java/com/appeal/api/portfolio/dto/template/TemplateDto.java
@@ -1,4 +1,0 @@
-package com.appeal.api.portfolio.dto.template;
-
-public interface TemplateDto {
-}
