--- appeal-api/src/main/java/com/appeal/api/portfolio/dto/template/templateone/TemplateOneCareerDto.java
@@ -1,13 +1,0 @@
-package com.appeal.api.portfolio.dto.template.templateone;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter @Setter
-public class TemplateOneCareerDto {
-
-    private String date;
-    private String title;
-    private String position;
-    private String name;
-}
