--- appeal-api/src/main/java/com/appeal/api/portfolio/dto/template/templateone/TemplateOneCertificateDto.java
@@ -1,11 +1,0 @@
-package com.appeal.api.portfolio.dto.template.templateone;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter @Setter
-public class TemplateOneCertificateDto {
-    private String date;
-    private String title;
-    private String origin;
-}
