--- appeal-api/src/main/java/com/appeal/api/portfolio/dto/template/templatetwo/TemplateTwoCareerDto.java
@@ -1,12 +1,0 @@
-package com.appeal.api.portfolio.dto.template.templatetwo;
-
-import lombok.*;
-
-@Getter @Setter @Builder @AllArgsConstructor @NoArgsConstructor
-public class TemplateTwoCareerDto {
-    private String title;
-    private String date;
-    private String position;
-    private String stack;
-    private String intro;
-}
