--- appeal-api/src/main/java/com/appeal/api/portfolio/dto/template/templatetwo/TemplateTwoProjectFileDto.java
@@ -1,13 +1,0 @@
-package com.appeal.api.portfolio.dto.template.templatetwo;
-
-import lombok.*;
-import org.springframework.web.multipart.MultipartFile;
-
-@Getter @Setter
-public class TemplateTwoProjectFileDto {
-
-    private String name;
-    private String intro;
-    private String role;
-    private MultipartFile thumbnail;
-}
