--- appeal-api/src/main/java/com/appeal/api/portfolio/repository/TemplateOneRepository.java
@@ -1,7 +1,0 @@
-package com.appeal.api.portfolio.repository;
-
-import com.appeal.api.portfolio.domain.templateone.TemplateOne;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TemplateOneRepository extends JpaRepository<TemplateOne, Long> {
-}
