+++ appeal-api/src/main/java/com/appeal/api/portfolio/service/PortfolioService.java
@@ -4,7 +4,7 @@ import com.appeal.api.member.domain.Member;
 import com.appeal.api.member.dto.MemberSession;
 import com.appeal.api.member.repository.MemberRepository;
 import com.appeal.api.portfolio.domain.Portfolio;
-import com.appeal.api.portfolio.dto.portfolio.PortfolioDto;
+import com.appeal.api.portfolio.dto.PortfolioDto;
 import com.appeal.api.portfolio.repository.PortfolioRepository;
 import com.appeal.exception.notfound.NotFoundMemberException;
 import lombok.RequiredArgsConstructor;
@@ -12,7 +12,6 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 @RequiredArgsConstructor
 @Service
