+++ appeal-api/src/main/java/com/appeal/api/template/controller/TemplateOneController.java
@@ -1,0 +1,14 @@
+package com.appeal.api.template.controller;
+
+import com.appeal.api.template.dto.templateone.TemplateOneFileDto;
+import com.appeal.api.template.service.TemplateOneService;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/templateone")
+public class TemplateOneController extends TemplateController<TemplateOneFileDto, TemplateOneService> {
+    public TemplateOneController(TemplateOneService templateService) {
+        super(templateService);
+    }
+}
