+++ appeal-api/src/main/java/com/appeal/api/template/dto/TemplateDto.java
@@ -1,0 +1,4 @@
+package com.appeal.api.template.dto;
+
+public interface TemplateDto {
+}
