+++ appeal-api/src/main/java/com/appeal/api/template/dto/templateone/TemplateOneCareerDto.java
@@ -1,0 +1,13 @@
+package com.appeal.api.template.dto.templateone;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter @Setter
+public class TemplateOneCareerDto {
+
+    private String date;
+    private String title;
+    private String position;
+    private String name;
+}
