+++ appeal-api/src/main/java/com/appeal/api/template/dto/templateone/TemplateOneCertificateDto.java
@@ -1,0 +1,11 @@
+package com.appeal.api.template.dto.templateone;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter @Setter
+public class TemplateOneCertificateDto {
+    private String date;
+    private String title;
+    private String origin;
+}
