+++ appeal-api/src/main/java/com/appeal/api/template/dto/templatetwo/TemplateTwoCareerDto.java
@@ -1,0 +1,12 @@
+package com.appeal.api.template.dto.templatetwo;
+
+import lombok.*;
+
+@Getter @Setter @Builder @AllArgsConstructor @NoArgsConstructor
+public class TemplateTwoCareerDto {
+    private String title;
+    private String date;
+    private String position;
+    private String stack;
+    private String intro;
+}
