+++ appeal-api/src/main/java/com/appeal/api/template/dto/templatetwo/TemplateTwoProjectFileDto.java
@@ -1,0 +1,13 @@
+package com.appeal.api.template.dto.templatetwo;
+
+import lombok.*;
+import org.springframework.web.multipart.MultipartFile;
+
+@Getter @Setter
+public class TemplateTwoProjectFileDto {
+
+    private String name;
+    private String intro;
+    private String role;
+    private MultipartFile thumbnail;
+}
