+++ appeal-api/src/main/java/com/appeal/api/template/repository/TemplateOneRepository.java
@@ -1,0 +1,7 @@
+package com.appeal.api.template.repository;
+
+import com.appeal.api.template.domain.templateone.TemplateOne;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TemplateOneRepository extends JpaRepository<TemplateOne, Long> {
+}
