+++ appeal-api/src/main/java/com/appeal/api/template/service/TemplateService.java
@@ -1,0 +1,10 @@
+package com.appeal.api.template.service;
+
+import com.appeal.api.member.dto.MemberSession;
+import com.appeal.api.template.dto.TemplateDto;
+
+public interface TemplateService {
+    Long createTemplate(MemberSession session, TemplateDto dto);
+
+    TemplateDto getTemplateById(Long id);
+}
